refactor: SectionsController, Service 삭제
