+++ src/main/java/nextstep/subway/ui/LineController.java
@@ -3,6 +3,8 @@ package nextstep.subway.ui;
 import nextstep.subway.application.dto.LineRequest;
 import nextstep.subway.application.dto.LineResponse;
 import nextstep.subway.application.LineService;
+import nextstep.subway.application.dto.SectionRequest;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -42,6 +44,17 @@ public class LineController {
     @DeleteMapping("/lines/{id}")
     public ResponseEntity deleteLine(@PathVariable Long id) {
         lineService.deleteLine(id);
+        return ResponseEntity.noContent().build();
+    }
+    @PostMapping("/lines/{lineId}/sections")
+    public ResponseEntity<Void> createSection(@PathVariable Long lineId, @RequestBody SectionRequest sectionRequest) {
+        lineService.saveSection(lineId, sectionRequest);
+        return ResponseEntity.status(HttpStatus.CREATED).build();
+    }
+
+    @DeleteMapping("/lines/{lineId}/sections")
+    public ResponseEntity deleteSection(@RequestParam Long stationId, @PathVariable Long lineId) {
+        lineService.deleteSection(stationId, lineId);
         return ResponseEntity.noContent().build();
     }
 }
